+++ src/GridCal/Engine/Simulations/OPF/ac_opf.py
@@ -51,6 +51,25 @@ def get_objective_function(Pg, Pb, LSlack, FSlack1, FSlack2,
     return f_obj
 
 
+def set_fix_generation(problem, Pg, P_fix, enabled_for_dispatch):
+    """
+    Set the generation fixed at the non dispatchable generators
+    :param problem: LP problem instance
+    :param Pg: Array of generation variables
+    :param P_fix: Array of fixed generation values
+    :param enabled_for_dispatch: array of "enables" for dispatching generators
+    :return: Nothing
+    """
+
+    idx = np.where(enabled_for_dispatch == False)[0]
+
+    lpAddRestrictions2(problem=problem,
+                       lhs=Pg[idx],
+                       rhs=P_fix[idx],
+                       name='fixed_generation',
+                       op='=')
+
+
 def get_power_injections(C_bus_gen, Pg, C_bus_bat, Pb, C_bus_load, PlSlack, QlSlack, Pl, Ql):
     """
     Create the power injections per bus
@@ -234,6 +253,8 @@ class AcOpf(Opf):
         Pg_max = numerical_circuit.generator_pmax / Sbase
         Pg_min = numerical_circuit.generator_pmin / Sbase
         cost_g = numerical_circuit.generator_cost
+        P_fix = numerical_circuit.generator_power / Sbase
+        enabled_for_dispatch = numerical_circuit.generator_dispatchable
 
         # load
         Pl = (numerical_circuit.load_active * numerical_circuit.load_power.real) / Sbase
@@ -262,6 +283,9 @@ class AcOpf(Opf):
         # add the objective function
         problem += get_objective_function(Pg, Pb, load_slack, branch_rating_slack1, branch_rating_slack2,
                                           cost_g, cost_b, cost_l, cost_br)
+
+        # set the fixed generation values
+        set_fix_generation(problem=problem, Pg=Pg, P_fix=P_fix, enabled_for_dispatch=enabled_for_dispatch)
 
         # compute the power injections per node
         P, Q = get_power_injections(C_bus_gen=numerical_circuit.C_gen_bus, Pg=Pg,
@@ -314,6 +338,9 @@ if __name__ == '__main__':
         # fname = '/home/santi/Documentos/GitHub/GridCal/Grids_and_profiles/grids/IEEE39_1W.gridcal'
 
         main_circuit = FileOpen(fname).open()
+
+        main_circuit.buses[3].controlled_generators[0].enabled_dispatch = False
+
         numerical_circuit_ = main_circuit.compile()
         problem = AcOpf(numerical_circuit=numerical_circuit_)