+++ src/GridCal/Engine/Simulations/OPF/dc_opf_ts.py
@@ -25,12 +25,10 @@ from GridCal.Engine.Simulations.OPF.opf_templates import OpfTimeSeries
 from GridCal.ThirdParty.pulp import *
 
 
-def add_objective_function(problem: LpProblem,
-                           Pg, Pb, LSlack, FSlack1, FSlack2,
+def get_objective_function(Pg, Pb, LSlack, FSlack1, FSlack2,
                            cost_g, cost_b, cost_l, cost_br):
     """
     Add the objective function to the problem
-    :param problem: LpProblem instance
     :param Pg: generator LpVars (ng, nt)
     :param Pb: batteries LpVars (nb, nt)
     :param LSlack: Load slack LpVars (nl, nt)
@@ -51,7 +49,27 @@ def add_objective_function(problem: LpProblem,
 
     f_obj += (cost_br * (FSlack1 + FSlack2)).sum()
 
-    problem += f_obj
+    return f_obj
+
+
+def set_fix_generation(problem, Pg, P_profile, enabled_for_dispatch):
+    """
+    Set the generation fixed at the non dispatchable generators
+    :param problem: LP problem instance
+    :param Pg: Array of generation variables
+    :param P_profile: Array of fixed generation values
+    :param enabled_for_dispatch: array of "enables" for dispatching generators
+    :return: Nothing
+    """
+
+    idx = np.where(enabled_for_dispatch == False)[0]
+
+    lpAddRestrictions2(problem=problem,
+                       lhs=Pg[idx, :],
+                       rhs=P_profile[idx, :],
+                       # Fmax + FSlack2
+                       name='fixed_generation',
+                       op='=')
 
 
 def get_power_injections(C_bus_gen, Pg, C_bus_bat, Pb, C_bus_load, LSlack, Pl):
@@ -262,7 +280,9 @@ class OpfDcTimeSeries(OpfTimeSeries):
         # generator
         Pg_max = numerical_circuit.generator_pmax / Sbase
         Pg_min = numerical_circuit.generator_pmin / Sbase
+        P_profile = numerical_circuit.generator_power_profile[a:b, :].transpose() / Sbase
         cost_g = numerical_circuit.generator_cost_profile[a:b, :].transpose()
+        enabled_for_dispatch = numerical_circuit.generator_dispatchable
 
         # load
         Pl = (numerical_circuit.load_active_prof[a:b, :] * numerical_circuit.load_power_profile.real[a:b, :]).transpose() / Sbase
@@ -294,9 +314,14 @@ class OpfDcTimeSeries(OpfTimeSeries):
         problem = LpProblem(name='DC_OPF_Time_Series')
 
         # add the objective function
-        add_objective_function(problem, Pg, Pb, load_slack, branch_rating_slack1, branch_rating_slack2,
-                               cost_g, cost_b, cost_l, cost_br)
+        problem += get_objective_function(Pg, Pb, load_slack, branch_rating_slack1, branch_rating_slack2,
+                                          cost_g, cost_b, cost_l, cost_br)
 
+        # set the fixed generation values
+        set_fix_generation(problem=problem, Pg=Pg, P_profile=P_profile,
+                           enabled_for_dispatch=enabled_for_dispatch)
+
+        # compute the power injections
         P = get_power_injections(C_bus_gen=numerical_circuit.C_gen_bus,
                                  Pg=Pg,
                                  C_bus_bat=numerical_circuit.C_batt_bus,
@@ -305,6 +330,7 @@ class OpfDcTimeSeries(OpfTimeSeries):
                                  LSlack=load_slack,
                                  Pl=Pl)
 
+        # set the nodal restrictions
         nodal_restrictions = add_dc_nodal_power_balance(numerical_circuit, problem, theta, P,
                                                         start_=self.start_idx, end_=self.end_idx)
 
@@ -334,53 +360,46 @@ class OpfDcTimeSeries(OpfTimeSeries):
 
 if __name__ == '__main__':
 
-        from GridCal.Engine import *
+    from GridCal.Engine import *
 
-        # fname = '/home/santi/Documentos/GitHub/GridCal/Grids_and_profiles/grids/Lynn 5 Bus pv.gridcal'
-        fname = '/home/santi/Documentos/GitHub/GridCal/Grids_and_profiles/grids/IEEE39_1W.gridcal'
+    fname = '/home/santi/Documentos/GitHub/GridCal/Grids_and_profiles/grids/Lynn 5 Bus pv.gridcal'
+    # fname = '/home/santi/Documentos/GitHub/GridCal/Grids_and_profiles/grids/IEEE39_1W.gridcal'
 
-        main_circuit = FileOpen(fname).open()
-        # numerical_circuit_ = main_circuit.compile()
-        # problem = OpfDcTimeSeries(numerical_circuit=numerical_circuit_, start_idx=5, end_idx=5 + 5 * 24)
-        #
-        # print('Solving...')
-        # status = problem.solve()
+    main_circuit = FileOpen(fname).open()
 
-        # get the power flow options from the GUI
-        solver = SolverType.DC_OPF
-        mip_solver = MIPSolvers.CBC
-        grouping = TimeGrouping.Daily
-        pf_options = PowerFlowOptions()
+    # main_circuit.buses[3].controlled_generators[0].enabled_dispatch = False
 
-        options = OptimalPowerFlowOptions(solver=solver,
-                                          grouping=grouping,
-                                          mip_solver=mip_solver,
-                                          power_flow_options=pf_options)
+    # get the power flow options from the GUI
+    solver = SolverType.DC_OPF
+    mip_solver = MIPSolvers.CBC
+    grouping = TimeGrouping.NoGrouping
+    pf_options = PowerFlowOptions()
 
-        start = 0
-        end = len(main_circuit.time_profile)
+    options = OptimalPowerFlowOptions(solver=solver,
+                                      grouping=grouping,
+                                      mip_solver=mip_solver,
+                                      power_flow_options=pf_options)
 
-        # create the OPF time series instance
-        # if non_sequential:
-        optimal_power_flow_time_series = OptimalPowerFlowTimeSeries(grid=main_circuit,
-                                                                    options=options,
-                                                                    start_=start,
-                                                                    end_=end)
+    start = 0
+    end = len(main_circuit.time_profile)
 
-        optimal_power_flow_time_series.run()
+    # create the OPF time series instance
+    # if non_sequential:
+    optimal_power_flow_time_series = OptimalPowerFlowTimeSeries(grid=main_circuit,
+                                                                options=options,
+                                                                start_=start,
+                                                                end_=end)
 
-        # print("Status:", status)
+    optimal_power_flow_time_series.run()
 
-        # v = problem.get_voltage()
-        # print('Angles\n', np.angle(v))
-        #
-        # l = problem.get_loading()
-        # print('Branch loading\n', l)
-        #
-        # g = problem.get_generator_power()
-        # print('Gen power\n', g)
-        #
-        # pr = problem.get_shadow_prices()
-        # print('Nodal prices \n', pr)
-        #
-        # pass
+    v = optimal_power_flow_time_series.results.voltage
+    print('Angles\n', np.angle(v))
+
+    l = optimal_power_flow_time_series.results.loading
+    print('Branch loading\n', l)
+
+    g = optimal_power_flow_time_series.results.controlled_generator_power
+    print('Gen power\n', g)
+
+    pr = optimal_power_flow_time_series.results.shadow_prices
+    print('Nodal prices \n', pr)