+++ src/GridCal/Gui/GuiFunctions.py
@@ -541,12 +541,12 @@ class ObjectsModel(QtCore.QAbstractTableModel):
                 delegate = ComboDelegate(self.parent, [True, False], ['True', 'False'])
                 F(i, delegate)
 
-            if tpe is BranchType:
+            elif tpe is BranchType:
                 conv = BranchTypeConverter(None)
                 delegate = ComboDelegate(self.parent, conv.values, conv.options)
                 F(i, delegate)
 
-            if tpe is BranchTemplate:
+            elif tpe is BranchTemplate:
                 F(i, None)
 
             elif tpe is float:
@@ -688,6 +688,8 @@ class ObjectsModel(QtCore.QAbstractTableModel):
                     setattr(self.objects[obj_idx], self.attributes[attr_idx], value)
             else:
                 pass  # the column cannot be edited
+
+        return True
 
     def attr_taken(self, attr, val):
         """